Se creo el edit y el destroy, hay que mover el destroy de lugar

+++ database/migrations/2023_06_14_000827_create_posts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('price');
+            $table->bigInteger('price');
             $table->integer('stock');
             $table->text('image');
             $table->integer('category_id');
